pyproject.toml
[project]
name = "classhelper"
version = "0.6.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "async-tkinter-loop>=0.9.3",
    "kaleido==0.2.1",
    "pillow>=11.2.1",
    "plotly>=6.0.1",
    "pyinstaller>=6.13.0",
    "pyinstaller-hooks-contrib>=2025.4",
    "simpleaudio>=1.0.4",
]
